backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
bcrypt==4.1.2
python-multipart==0.0.12
email-validator==2.3.0
alembic==1.13.3
python-dotenv==1.0.1
httpx==0.27.2
